gsk: Rename ngl to gl
For now, this is just doing the code reshuffling. The environment
variable still expects you to set
GSK_RENDERER=ngl

gsk_ngl_renderer_new is still around, deprecated, and returns
a renderer that fails to realize.

The gsk/gl/gskglrenderer.h is still installed, but including it
triggers a deprecation warning.